Cleaned up schedulers and webhooks
