+++ plugins/pakkasmarja-berries-utils/package.json
@@ -1,0 +1,19 @@
+{
+  "name": "pakkasmarja-berries-utils",
+  "version": "0.0.1",
+  "description": "Utilities for Pakkasmarja Berries",
+  "main": "index.js",
+  "plugin": {
+    "provides": [
+      "pakkasmarja-berries-utils"
+    ],
+    "consumes": [
+      "logger",
+      "pakkasmarja-berries-wordpress",
+      "pakkasmarja-berries-models",
+      "pakkasmarja-berries-user-management",
+      "pakkasmarja-berries-push-notifications",
+      "shady-messages"
+    ]
+  }
+}
