+++ plugins/pakkasmarja-berries-webhooks/package.json
@@ -10,10 +10,8 @@
     "consumes": [
       "logger",
       "pakkasmarja-berries-models",
-      "pakkasmarja-berries-user-management",
       "pakkasmarja-berries-wordpress",
-      "pakkasmarja-berries-push-notifications",
-      "shady-messages"
+      "pakkasmarja-berries-utils"
     ]
   }
 }
